New README, and a refactor to support multiple sources within an extractor, and support multiple extractors in the future
