+++ typescript/src/cli.ts
@@ -29,7 +29,7 @@ const run = async function(cliArgs: any): Promise<void> {
   const start = new Date().getTime();
 
   log('Parsing settings');
-  // CLI settings
+  // Parse CLI settings
   let jiraConfigPath: string = cliArgs.i ? cliArgs.i : defaultYamlPath;
   let isLegacyYaml: boolean = cliArgs.l ? true : false;
   let outputPath: string = cliArgs.o ? cliArgs.o : defaultOutputPath;
@@ -37,7 +37,8 @@ const run = async function(cliArgs: any): Promise<void> {
   if (outputType !== 'CSV') {
     throw new Error('Only CSV is currently supported. JSON support coming soon');
   }
-  // YAML settings
+
+  // Parse YAML settings
   let settings: any  = {};
   try {
     let yamlConfig = safeLoad(fs.readFileSync(jiraConfigPath, 'utf8'));
@@ -50,7 +51,7 @@ const run = async function(cliArgs: any): Promise<void> {
   const jiraSettings = new JiraSettings(settings, 'yaml');
   console.log('Successfully parsed settings');
   
-  // import data
+  // Import data
   log('Beginning extraction process');
   const jiraExtractor = new JiraExtractor(jiraSettings);
   try {
@@ -60,7 +61,7 @@ const run = async function(cliArgs: any): Promise<void> {
     throw e;
   }
 
-  //export data
+  // Export data
   let data: string;
   if (outputType === 'CSV') {
     data = jiraExtractor.toCSV();
@@ -84,3 +85,4 @@ const run = async function(cliArgs: any): Promise<void> {
     console.log(e);
   }
 }(getArgs()));
+
