+++ typescript/src/lib/extractors/extractors.ts
@@ -1,0 +1,1 @@
+export * from './jira/extractor';
