+++ typescript/src/lib/extractors/jira/api-adapter/components/attribute-parser.ts
