+++ typescript/src/lib/extractors/jira/api-adapter/components/query-builder.ts
