+++ typescript/src/lib/extractors/jira/api-adapter/components/staging-parser.ts
