+++ typescript/src/lib/extractors/jira/api-adapter/main.ts
@@ -2,8 +2,8 @@ import { IIssueList, IIssue } from '../models';
 import { IJiraSettings } from '../settings';
 import { buildJiraQueryUrl } from './components/query-builder';
 import { getAttributes } from './components/attribute-parser';
-import { IWorkItem, WorkItem } from '../../core/work-item';
-import { getJsonFromUrl, getHeaders } from '../../core/http';
+import { IWorkItem, WorkItem } from '../../../core/work-item';
+import { getJsonFromUrl, getHeaders } from '../../../core/http';
 import { getStagingDates } from './components/staging-parser';
 
 const getIssues = async function(query: string, username: string, password: string): Promise<IIssue[]> {
@@ -29,23 +29,28 @@ const createWorkItem = (issue: IIssue, settings: IJiraSettings): IWorkItem => {
 };
 
 const getWorkItemsBatch = async function(start: number, batchSize: number, settings: IJiraSettings): Promise<IWorkItem[]> {
-  const url = buildJiraQueryUrl(settings.ApiUrl, settings.Criteria.Projects, settings.Criteria.IssueTypes, settings.Criteria.Filters, start, batchSize);
+  const url = buildJiraQueryUrl(
+    settings.ApiUrl, 
+    settings.Criteria.Projects, 
+    settings.Criteria.IssueTypes, 
+    settings.Criteria.Filters, 
+    start, 
+    batchSize
+  );
   const issues = await getIssues(url, settings.Connection.Username, settings.Connection.Password);
   const workItems = issues.map(issue => createWorkItem(issue, settings));
   return workItems;
 };
 
 const getAllWorkItemsFromJiraApi = async function(settings: IJiraSettings, resultsPerBatch = 25): Promise<IWorkItem[]> {
-  let metadata: any;
-  // try {
-    metadata = await getJsonFromUrl(
+  const metadata = await getJsonFromUrl(
       buildJiraQueryUrl(
         settings.ApiUrl, 
         settings.Criteria.Projects, 
         settings.Criteria.IssueTypes, 
         settings.Criteria.Filters), 
       getHeaders(settings.Connection.Username, settings.Connection.Password)
-    );
+  );
 
   const totalJiras: number = metadata.total; 
   const batchSize: number = resultsPerBatch;
@@ -53,13 +58,16 @@ const getAllWorkItemsFromJiraApi = async function(settings: IJiraSettings, resul
 
   console.log(`Connection successful. ${totalJiras} issues detected.`);
 
-  let allWorkItems: IWorkItem[] = [];
+  const allWorkItems: IWorkItem[] = [];
   for  (let i = 0; i < totalBatches; i++) {
+    //log
     const rangeLower: number = i * batchSize;
     const rangeUpper: number = Math.min((i * batchSize) + batchSize - 1, totalJiras);
     console.log(`Extracting batch ${i + 1}/${totalBatches} -- Isssue(s) ${rangeLower}-${rangeUpper} out of ${totalJiras}`);
-    const workItemsBatch = await getWorkItemsBatch(i * batchSize, batchSize, settings);
-    allWorkItems.push(...workItemsBatch);
+
+    // extract
+    const workItemBatch = await getWorkItemsBatch(i * batchSize, batchSize, settings);
+    allWorkItems.push(...workItemBatch);
   }
   return allWorkItems;
 };
@@ -67,5 +75,4 @@ const getAllWorkItemsFromJiraApi = async function(settings: IJiraSettings, resul
 export { 
   getAllWorkItemsFromJiraApi,
   getWorkItemsBatch,
-  getStagingDates,
 };
