+++ typescript/src/lib/extractors/jira/models.ts
