+++ typescript/src/lib/extractors/jira/settings.ts
