+++ typescript/src/main.ts
@@ -1,2 +1,1 @@
-export { JiraExtractor } from './lib/jira/extractor';
-export { JiraSettings } from './lib/jira/settings';
+export * from './lib/extractors/extractors';
